Added percentages in stats.
Former-commit-id: 66e0c2e8e3df4f0c5f61708c8f416135e1e05286

+++ app/models.py
@@ -121,7 +121,7 @@ class Shuffle(models.Model):
 class Stats(models.Model):
     user = models.ForeignKey(User)
     track = models.ForeignKey(Track)
-    listeningPercentage = models.IntegerField(null=True)
+    listeningPercentage = models.FloatField(null=True)
     playCounter = models.IntegerField(default=0)
